# public/openapi.yaml
openapi: 3.1.1
info:
  title: "Cars API Documentation"
  description: |
    # A modern REST API built with Next.js

    Esta API proporciona las siguientes funcionalidades:

    ## **Autenticación de usuarios**
    - Registro de usuarios
    - Login de usuarios
    - Verificar estado de autenticación

    ## **Coches (Cars)**
    - Crear coches
    - Consultar coches
    - Actualizar coches
    - Eliminar coches
    
    ## **Imágenes**
    - Subir imágenes en cualquier formato (`image/*`)
    - Consultar imágenes subidas y sus metadatos
    - Actualizar imágenes subidas
    - Eliminar imágenes subidas

    > **NOTA:**
    > Las siguientes propiedades deben tener **valores únicos**: `brand` y `model` de un coche, así como el `email` de un usuario.
    >
    > - Si al introducir o modificar un coche obtienes un **código 409**, es porque ya existe un coche con el mismo `brand` o `model`.
    > - Si al registrar un usuario obtienes un **código 409**, es porque ya existe un usuario con el mismo `email`.

    Para más información, visita [nuestro repositorio](https://github.com/jamj2000/nxapi-shop).
  version: "1.0.0"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserCreate:
      type: object
      required:
        - email
        - password
        - fullName
      properties:
        email:
          type: string
          description: A valid email (unique)
          nullable: false
        password:
          type: string
        fullName:
          type: string

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: A valid email
          nullable: false
        password:
          type: string
          description: The password for this user
          nullable: false

    Car:
      type: object
      required:
        - id
        - brand
        - model
        - year
        - price
        - description
        - images
      properties:
        id:
          type: string
          description: Car ID
          uniqueItems: true
          example: 1
        brand:
          type: string
          description: Car Brand
          uniqueItems: true
          example: Tesla
        price:
          type: number
          description: Car price
          example: 45000
        description:
          type: string
          description: Car description
          default: null
          example: Coche eléctrico en perfecto estado.
        model:
          type: string
          description: Car Model
          uniqueItems: true
          example: model_3
        year:
          type: number
          description: Car year
          default: 2024
          example: 2024
        images:
          type: array
          items:
            type: string

    CarCreate:
      type: object
      required:
        - brand
        - price
        - description
        - model
        - year
        - images
      properties:
        brand:
          type: string
          description: Car brand (unique)
          nullable: false
          minLength: 1
        price:
          type: number
        description:
          type: string
        model:
          type: string
          description: Car model (unique)
          nullable: false
          minLength: 1
        year:
          type: number
        images:
          type: array
          items:
            type: string

    CarUpdate:
      type: object
      properties:
        brand:
          type: string
          description: Car brand (unique)
          nullable: false
          minLength: 1
        price:
          type: number
        description:
          type: string
        model:
          type: string
          description: Car model (unique)
          nullable: false
          minLength: 1
        year:
          type: number
        images:
          type: array
          items:
            type: string

    ImageInfo:
      type: object
      properties:
        asset_id:
          type: string
          nullable: false
          example: 9dbbd654a6986807fc5329032b19877d
        public_id:
          type: string
          nullable: false
          example: file_gvfsgc
        format:
          type: string
          nullable: false
          description: The format of the image (webp, jpg, png, etc)
          example: webp
        resource_type:
          type: string
          nullable: false
          description: The type of the resource (image, video, etc)
          example: image
        bytes:
          type: number
          nullable: false
          description: The size of the image in bytes
          example: 26468
        width:
          type: number
          nullable: false
          description: The width of the image in pixels
          example: 852
        height:
          type: number
          nullable: false
          description: The height of the image in pixels
          example: 852
        asset_folder:
          type: string
          nullable: false
          description: The folder where the image is stored
          example: escuela_coches
        secure_url:
          type: string
          format: uri
          nullable: false
          description: The secure URL of the image
          example: https://res.cloudinary.com/demo/image/upload/v1/coche.webp

paths:

  /api/auth/register:
    post:
      summary: Register
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Registration successful
        "400":
          description: Missing required fields
        "409":
          description: User already exists

  /api/auth/login:
    post:
      summary: Login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "201":
          description: Login successful
        "400":
          description: Missing email or password
        "401":
          description: Invalid credentials
        "404":
          description: User not found

  /api/auth/check-status:
    get:
      summary: Check authentication status
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Check successful
        "401":
          description: Unauthorized / Invalid token
        "404":
          description: User not found

  /api/cars:
    post:
      summary: Create new car
      tags:
        - Cars
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarCreate"
      responses:
        "201":
          description: Car was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "400":
          description: Missing data
        "401":
          description: Unauthorized. Token expired or invalid.
        "409":
          description: Car brand or model already exists

    get:
      summary: List cars
      tags:
        - Cars
      parameters:
        - name: limit
          in: query
          required: true
          description: How many rows do you need
          schema:
            type: number
            default: 10
        - name: offset
          in: query
          required: true
          description: How many rows do you want to skip
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: Car list
        "500":
          description: Internal server error

  /api/cars/all/{term}:
    get:
      summary: Search cars
      tags:
        - Cars
      parameters:
        - name: term
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
        "500":
          description: Internal server error

  /api/cars/{id}:
    get:
      summary: Find car by ID
      tags:
        - Cars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Car found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "500":
          description: Internal server error

    patch:
      summary: Update car
      tags:
        - Cars
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarUpdate"
      responses:
        "200":
          description: Car updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "401":
          description: Unauthorized. Token expired or invalid.
        "409":
          description: Car brand or model already exists
    delete:
      summary: Delete car
      tags:
        - Cars
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Car deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "401":
          description: Unauthorized. Token expired or invalid.
        "500":
          description: Internal server error

  /api/images:
    post:
      summary: Upload car image
      tags:
        - Files - Get and Upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                carId:
                  type: integer
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageInfo"
        "401":
          description: Unauthorized. Token expired or invalid.
        "500":
          description: Internal server error
    get:
      summary: List images
      tags:
        - Files - Get and Upload
      parameters:
        - name: limit
          in: query
          required: true
          description: How many rows do you need
          schema:
            type: number
            default: 10
      responses:
        "200":
          description: OK
        "500":
          description: Internal server error  

  /api/images/{imageName}:
    get:
      summary: Get car image
      tags:
        - Files - Get and Upload
      parameters:
        - name: imageName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Image not found
    delete:
      summary: Delete car image
      tags:
        - Files - Get and Upload
      security:
        - bearerAuth: []
      parameters:
        - name: imageName
          in: path
          required: true
          schema:
            type: string   
      responses:
        "200":
          description: Image deleted successfully
        "401":
          description: Unauthorized. Token expired or invalid.
        "500":
          description: Internal server error